add initial sns subscription implementation (#1)
Initial AWS ToolKit app and block implementation. The first block is
called `subscribeSNSTopic`, which as the name suggests, subscribes to
the SNS topic, confirms the subscription and emits received messages as
events.

Since subscription confirmation is async, a `kv` store was used. When a
`SubscriptionConfirmation` event is received, we confirm it and trigger
an additional sync to mark block as ready.

Additional
[sns-validator](https://github.com/aws/aws-js-sns-message-validator)
package was used to check that requests originate from AWS. While it's
older, it's working as intended.

Lastly, `subscriptionArn` is exposed via signals and on tier-down, the
subscription is remotely deleted.



https://github.com/user-attachments/assets/c84dd4c7-63af-4937-91ef-10535bac31a8